back/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
whitenoise
gunicorn
psycopg2-binary
dj-database-url
